backend/requirements.txt
fastapi
uvicorn
python-jose
pydantic
joblib
pandas
scikit-learn
